requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
audioop-lts==0.2.2
audioread==3.1.0
beautifulsoup4==4.14.2
blinker==1.9.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
chromaprint==0.5
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
decorator==5.2.1
fastapi==0.121.2
Flask==3.1.2
gunicorn==23.0.0
h11==0.16.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.46.0
lxml==6.0.2
MarkupSafe==3.0.3
msgpack==1.1.2
mysql-connector-python==9.5.0
numba==0.63.1
numpy==2.3.5
packaging==25.0
platformdirs==4.5.1
pooch==1.8.2
pyacoustid==1.3.0
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
PyMySQL==1.1.2
python-multipart==0.0.20
requests==2.32.5
scikit-learn==1.8.0
scipy==1.16.3
sniffio==1.3.1
sounddevice==0.5.3
soundfile==0.13.1
soupsieve==2.8
soxr==1.0.0
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
starlette==0.49.3
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
Werkzeug==3.1.4
